Move spacetime modules to B_granule_motion
Renamed and relocated B_medium.py and B_wave_engine.py to the B_granule_motion experiment directory for better modularity. Updated imports in related files to reflect new paths and standardized launcher/instrumentation file naming.
